Add commands for viewing and uploading build templates.
This allows you to see which templates are installed on a given project, as well as clone a template to a new Team Project.

Examples:
List all templates for a project:
tfx build templates list <projectName>

Get the JSON for a specific template:
tfx build templates list <projectName> --id <templateId>

Save the JSON output to a file. Manually tweak certain values if needed. Then upload the template to a new project:
tfx build templates upload <otherProject> path/to/template.json

The upload command can also be used to update templates that already exist.